Move sources to src/folder
