--- BeagleBone_black/main.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-
-int main() {
-  std::cout << "Starting BeagleBone Black..." << std::endl;
-  return 0;
-}
