+++ BeagleBone_black/src/main.cpp
@@ -1,0 +1,6 @@
+#include <iostream>
+
+int main() {
+  std::cout << "Starting BeagleBone Black..." << std::endl;
+  return 0;
+}
